+++ JavaSource/org/unitime/timetable/test/BatchStudentSectioningLoader.java
@@ -200,8 +200,8 @@ public class BatchStudentSectioningLoader extends StudentSectioningLoader {
                     (room instanceof Room? ((Room)room).getBuilding().getUniqueId() : null),
                     0,
                     room.getCapacity().intValue(),
-                    room.getCoordinateX().intValue(),
-                    room.getCoordinateY().intValue(),
+                    room.getCoordinateX(),
+                    room.getCoordinateY(),
                     room.isIgnoreTooFar().booleanValue(),
                     null);
             rooms.addElement(roomLocation);
